+++ internal/manager/feature_gates.go
@@ -22,6 +22,10 @@ const (
 	// objects like Ingress instead of creating a route per path.
 	combinedRoutesFeature = "CombinedRoutes"
 
+	// ingressClassParametersFeature is the name of the feature-gate for enabling/disabling
+	// IngressClassParameters CRD support.
+	ingressClassParametersFeature = "IngressClassParameters"
+
 	// featureGatesDocsURL provides a link to the documentation for feature gates in the KIC repository.
 	featureGatesDocsURL = "https://github.com/Kong/kubernetes-ingress-controller/blob/main/FEATURE_GATES.md"
 )
@@ -51,8 +55,9 @@ func setupFeatureGates(setupLog logr.Logger, c *Config) (map[string]bool, error)
 // NOTE: if you're adding a new feature gate, it needs to be added here.
 func getFeatureGatesDefaults() map[string]bool {
 	return map[string]bool{
-		knativeFeature:        false,
-		gatewayFeature:        false,
-		combinedRoutesFeature: false,
+		knativeFeature:                false,
+		gatewayFeature:                false,
+		combinedRoutesFeature:         false,
+		ingressClassParametersFeature: false,
 	}
 }