+++ test/e2e/kuma_test.go
@@ -5,6 +5,7 @@ package e2e
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
 	"github.com/blang/semver/v4"
@@ -77,8 +78,24 @@ func TestDeployAllInOneDBLESSKuma(t *testing.T) {
 	t.Log("running ingress tests to verify all-in-one deployed ingress controller and proxy are functional")
 	deployIngress(ctx, t, env)
 	service, err := env.Cluster().Client().CoreV1().Services("default").Get(ctx, "httpbin", metav1.GetOptions{})
+	require.NoError(t, err)
+
+	t.Logf("service %#v", service)
+	if service.ObjectMeta.Annotations == nil {
+		service.ObjectMeta.Annotations = map[string]string{}
+	}
 	service.ObjectMeta.Annotations["ingress.kubernetes.io/service-upstream"] = "true"
-	service, err = env.Cluster().Client().CoreV1().Services("default").Update(ctx, service, metav1.UpdateOptions{})
+	_, err = env.Cluster().Client().CoreV1().Services("default").Update(ctx, service, metav1.UpdateOptions{})
+	require.NoError(t, err,
+		// dump the status of service if the error happens on updating service.
+		func() string {
+			service, err := env.Cluster().Client().CoreV1().Services("default").Get(ctx, "httpbin", metav1.GetOptions{})
+			if err != nil {
+				return fmt.Sprintf("failed to dump service, error %v", err)
+			}
+			return fmt.Sprintf("current status of service: %#v", service)
+		}(),
+	)
 	verifyIngress(ctx, t, env)
 }
 
@@ -145,11 +162,21 @@ func TestDeployAllInOnePostgresKuma(t *testing.T) {
 	service, err := env.Cluster().Client().CoreV1().Services("default").Get(ctx, "httpbin", metav1.GetOptions{})
 	require.NoError(t, err)
 
+	t.Logf("current status of service: %#v", service)
 	if service.ObjectMeta.Annotations == nil {
 		service.ObjectMeta.Annotations = map[string]string{}
 	}
 	service.ObjectMeta.Annotations["ingress.kubernetes.io/service-upstream"] = "true"
-	service, err = env.Cluster().Client().CoreV1().Services("default").Update(ctx, service, metav1.UpdateOptions{})
-	require.NoError(t, err)
+	_, err = env.Cluster().Client().CoreV1().Services("default").Update(ctx, service, metav1.UpdateOptions{})
+	require.NoError(t, err,
+		// dump the status of service if the error happens on updating service.
+		func() string {
+			service, err := env.Cluster().Client().CoreV1().Services("default").Get(ctx, "httpbin", metav1.GetOptions{})
+			if err != nil {
+				return fmt.Sprintf("failed to dump service, error %v", err)
+			}
+			return fmt.Sprintf("current status of service: %#v", service)
+		}(),
+	)
 	verifyIngress(ctx, t, env)
 }